feature(dao) modificaciones en los daos generic, mascota y microchip

+++ Main/src/dao/GenericDao.java
@@ -12,10 +12,10 @@ import java.util.List;
  */
 public interface GenericDao<T> {
     void crear(T entidad) throws Exception;
-    T leer(int id) throws Exception;
+    T leer(Long id) throws Exception;
     List<T> leerTodos() throws Exception;
     void actualizar(T entidad) throws Exception;
-    void eliminar(int id_dom) throws Exception;
+    void eliminar(Long id_dom) throws Exception;
     
     // void recuperar(int id) throws Exception;
 }
